+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autons/JewelAutonBlue.java
@@ -1,11 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Autons;
 
 
 import android.graphics.Color;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -13,6 +12,10 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.navigation.Position;
 import org.firstinspires.ftc.robotcore.external.navigation.Velocity;
+import org.firstinspires.ftc.teamcode.Utilities.Gyro;
+import org.firstinspires.ftc.teamcode.Utilities.Logging;
+import org.firstinspires.ftc.teamcode.Utilities.Robot;
+import org.firstinspires.ftc.teamcode.Utilities.RobotMap;
 
 @Autonomous(name = "Blue: Jewel Auton", group = "Sensor")
 public class JewelAutonBlue extends LinearOpMode {
