+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/DriveDistanceTest.java
@@ -1,12 +1,16 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Testing;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
+import org.firstinspires.ftc.teamcode.Utilities.Logging;
+import org.firstinspires.ftc.teamcode.Utilities.PID;
+import org.firstinspires.ftc.teamcode.Utilities.Robot;
+import org.firstinspires.ftc.teamcode.Utilities.RobotMap;
+
 @Autonomous(name = "Driving Distance test", group = "Test")
-public class DriveDistanceAuton extends LinearOpMode{
+public class DriveDistanceTest extends LinearOpMode{
 
     public void runOpMode(){
         Robot robot = new Robot(this.hardwareMap);
