+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/IMUTurning_Test.java
@@ -27,25 +27,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Testing;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
-import com.qualcomm.hardware.bosch.JustLoggingAccelerationIntegrator;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.robotcore.external.Func;
-import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.navigation.Acceleration;
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
-import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 import org.firstinspires.ftc.robotcore.external.navigation.Position;
 import org.firstinspires.ftc.robotcore.external.navigation.Velocity;
-
-import java.util.Arrays;
-import java.util.Locale;
+import org.firstinspires.ftc.teamcode.Utilities.Gyro;
+import org.firstinspires.ftc.teamcode.Utilities.Logging;
+import org.firstinspires.ftc.teamcode.Utilities.Robot;
+import org.firstinspires.ftc.teamcode.Utilities.RobotMap;
 
 @Autonomous(name = "Turn 90 test", group = "Sensor")
 public class IMUTurning_Test extends LinearOpMode
