+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/SensorREVColorDistance.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Testing;
 
 import android.app.Activity;
 import android.graphics.Color;
