+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/TickCounter.java
@@ -1,8 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Testing;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
+
+import org.firstinspires.ftc.teamcode.Utilities.Robot;
 
 @Autonomous(name = "Tick Counter test", group = "Test")
