+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utilities/Logging.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Utilities;
 
 import android.util.Log;
