+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utilities/RobotMap.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Utilities;
 
 public class RobotMap {
     public static final float GEAR_RATIO = 50;
