+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utilities/SensorColor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Utilities;
 
 import android.graphics.Color;
 import android.util.Log;
